Import Ballistic knife animations (1.6-D1)
Import Ballistic knife animations from CoD black ops for use with the
"Assassins Blade".
Added kill icon for the Assassin's Blade.
Checked whether a player is in third person or not in killcam, and reset
thirdperson on spawn.
Changed Assassin's weapon from bulletweapon to projectile weapon for
possible implementation of ballistic knives.